requirements.txt
Django==4.2.13
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
argon2-cffi==23.1.0
pytest-django==4.8.0
pytest==8.1.0
